feat: allow filtering flights from map page
