+++ src/lib/components/flight-filters/types.ts
@@ -1,8 +1,15 @@
 import type { CalendarDate } from '@internationalized/date';
 
-export type ToolbarFilters = {
+export type FlightFilters = {
   departureAirports: string[];
   arrivalAirports: string[];
   fromDate: CalendarDate | undefined;
   toDate: CalendarDate | undefined;
+};
+
+export const defaultFilters: FlightFilters = {
+  departureAirports: [],
+  arrivalAirports: [],
+  fromDate: undefined,
+  toDate: undefined,
 };
