+++ src/lib/components/map/index.ts
@@ -1,2 +1,3 @@
 export { default as Map } from './Map.svelte';
-export { default as Airports } from './Airports.svelte';
+export { default as AirportLayer } from './AirportLayer.svelte';
+export { default as FlightArcLayer } from './FlightArcLayer.svelte';
