+++ src/lib/state.svelte.ts
@@ -1,10 +1,6 @@
-import type { FlightData } from "./utils";
-
 export const openModalsState = $state({
   addFlight: false,
   listFlights: false,
   statistics: false,
   settings: false,
 });
-
-export const filteredFlightDataState = $state({flightData: [] as FlightData[]});
