+++ src/routes/+page.svelte
@@ -4,7 +4,12 @@
   import { toast } from 'svelte-sonner';
   import { ListFlightsModal, StatisticsModal } from '$lib/components/modals';
   import { Map } from '$lib/components/map';
-  import { filteredFlightDataState, openModalsState } from '$lib/stores.svelte';
+  import { openModalsState } from '$lib/state.svelte';
+  import { isAfter, isBefore } from 'date-fns';
+  import {
+    defaultFilters,
+    type FlightFilters,
+  } from '$lib/components/flight-filters/types';
 
   const rawFlights = trpc.flight.list.query();
 
@@ -15,9 +20,31 @@
     return prepareFlightData(data);
   });
 
-  $effect(() => {
-    filteredFlightDataState.flightData = flights;
-  })
+  let filters: FlightFilters = $state(defaultFilters);
+
+  const filteredFlights = $derived.by(() => {
+    return flights.filter((f) => {
+      if (
+        (filters.departureAirports.length &&
+          !filters.departureAirports.includes(f.from.ICAO)) ||
+        (filters.arrivalAirports.length &&
+          !filters.arrivalAirports.includes(f.to.ICAO))
+      ) {
+        return false;
+      } else if (
+        filters.fromDate &&
+        isBefore(f.date, filters.fromDate.toDate(f.date.timeZone ?? 'UTC'))
+      ) {
+        return false;
+      } else if (
+        filters.toDate &&
+        isAfter(f.date, filters.toDate.toDate(f.date.timeZone ?? 'UTC'))
+      ) {
+        return false;
+      }
+      return true;
+    });
+  });
 
   const invalidator = {
     onSuccess: () => {
@@ -40,9 +67,14 @@
 
 <ListFlightsModal
   bind:open={openModalsState.listFlights}
+  bind:filters
   {flights}
+  {filteredFlights}
   {deleteFlight}
 />
-<StatisticsModal bind:open={openModalsState.statistics} allFlights={filteredFlightDataState.flightData} />
+<StatisticsModal
+  bind:open={openModalsState.statistics}
+  allFlights={filteredFlights}
+/>
 
-<Map flights={filteredFlightDataState.flightData} />
+<Map bind:filters {flights} {filteredFlights} />
